requirements.txt
Flask
PyPDF2
faiss-cpu
flask
langchain
langchain-community
langchain-openai
markdown
numpy
scikit-learn
sentence-transformers
torch
tiktoken
transformers
